+++ src/lib/abstract-collection/abstract-collection.component.ts
@@ -1,0 +1,33 @@
+import {AbstractCollectionType} from '../form-type/abstract-collection-type';
+import {AbstractFormFieldComponent} from '../abstract-form-field/abstract-form-field.component';
+import {ModelHandler} from '../model-handler';
+import {FormArray} from '../abstraction';
+import {Directive, Input} from '@angular/core';
+
+// noinspection JSUnusedGlobalSymbols
+@Directive()
+export abstract class AbstractCollectionComponent extends AbstractFormFieldComponent<AbstractCollectionType<any, any>> {
+
+  @Input()
+  public mwElement: FormArray;
+
+  // noinspection JSUnusedGlobalSymbols
+  public removeEntry(index: number): void {
+    this.mwElement.removeAt(index);
+  }
+
+  // noinspection JSUnusedGlobalSymbols
+  public addEntry(): void {
+    const newControl = ModelHandler.buildSingleField(
+      this.mwFieldType.fieldInstance,
+      this.mwFieldType.builderInstance
+    );
+
+    if (!newControl) {
+      return;
+    }
+
+    newControl.initHandler.setIsInitialized(true);
+    this.mwElement.controls.push(newControl);
+  }
+}