+++ src/lib/form-type/abstract-collection-type.ts
@@ -1,4 +1,6 @@
 import {AbstractType} from './abstract-type';
+import {FormArray} from '../abstraction';
+import {ConstructorOf1} from '../types';
 
 export interface ICollectionTypeOptions<T> {
   entrySettings: T;
@@ -7,7 +9,10 @@ export interface ICollectionTypeOptions<T> {
 export abstract class AbstractCollectionType<TConfig, T> extends AbstractType<T & ICollectionTypeOptions<TConfig>> {
   public readonly fieldInstance: AbstractType<TConfig>;
 
-  constructor(public readonly childType: new(options: TConfig) => AbstractType<TConfig>,
+  public control: FormArray;
+
+  // noinspection TypeScriptAbstractClassConstructorCanBeMadeProtected
+  constructor(public readonly childType: ConstructorOf1<TConfig, AbstractType<TConfig>>,
               options: T & ICollectionTypeOptions<TConfig>) {
     super(options);
     this.fieldInstance = new this.childType(this.options.entrySettings);