+++ src/lib/form-type/abstract-group-type.ts
@@ -1,9 +1,10 @@
-import {AbstractType} from './abstract-type';
-import {FormModel} from '../form-builder/form-builder.component';
+import {AbstractType, FormModel} from './abstract-type';
+import {FormGroup} from '../abstraction';
 
 export interface IGroupTypeOptions {
   model: FormModel;
 }
 
 export abstract class AbstractGroupType<T extends IGroupTypeOptions> extends AbstractType<T> {
+  public control: FormGroup;
 }