+++ src/public-api.ts
@@ -2,13 +2,17 @@
  * Public API Surface of form-builder
  */
 
-export * from './lib/abstract-collection-form-field/abstract-collection-form-field.component';
+export * from './lib/abstract-collection/abstract-collection.component';
 export * from './lib/abstract-form-field/abstract-form-field.component';
-export * from './lib/abstract-group-form-field/abstract-group-form-field.component';
+export * from './lib/abstract-form-group/abstract-form-group.component';
+export * from './lib/abstract-layout/abstract-layout.component';
 export * from './lib/form-builder/form-builder.component';
-export * from './lib/form-field/form-field.component';
-export * from './lib/form-slot/form-slot.component';
+export * from './lib/form-field/form-field.directive';
+export * from './lib/form-slot/form-slot.directive';
 export * from './lib/form-builder.module';
-export * from './lib/form-type/abstract-type';
 export * from './lib/form-type/abstract-group-type';
 export * from './lib/form-type/abstract-collection-type';
+export * from './lib/form-type/abstract-layout-type';
+export * from './lib/form-type/abstract-type';
+
+export * from './lib/types';